+++ src/main/java/kernbeisser/Windows/ShoppingMask/ShoppingMaskUIController.java
@@ -3,7 +3,6 @@ package kernbeisser.Windows.ShoppingMask;
 import java.awt.*;
 import java.util.Objects;
 import javax.swing.*;
-import kernbeisser.CustomComponents.ShoppingTable.ShoppingCartController;
 import kernbeisser.DBEntities.*;
 import kernbeisser.Enums.MetricUnits;
 import kernbeisser.Enums.Mode;
@@ -14,6 +13,7 @@ import kernbeisser.Windows.MVC.IController;
 import kernbeisser.Windows.MVC.Linked;
 import kernbeisser.Windows.Pay.PayController;
 import kernbeisser.Windows.ShoppingMask.ArticleSelector.ArticleSelectorController;
+import kernbeisser.Windows.ShoppingMask.ShoppingTable.ShoppingCartController;
 import kernbeisser.Windows.Window;
 import kernbeisser.Windows.WindowImpl.SubWindow;
 import org.jetbrains.annotations.NotNull;
